Changed Endpoint properties to public, so they should be properly serialized as a health result #3402
